Make placeholders more easily deletable (if empty)

+++ frontend/src/plate/plugins/prohibit-deletion/prohibit-deletion.ts
@@ -2,6 +2,10 @@ import { AnyObject, createPluginFactory, findNode, insertElements } from '@udeco
 import { EditorFragmentDeletionOptions, EditorNormalizeOptions, TextUnit } from 'slate';
 import { ELEMENT_FOOTER, UNINTERACTIONABLE } from '@app/plate/plugins/element-types';
 import {
+  handleDeleteBackwardInPlaceholder,
+  handleDeleteForwardInPlaceholder,
+} from '@app/plate/plugins/prohibit-deletion/helpers';
+import {
   handleDeleteBackwardIntoUnchangeable,
   handleDeleteForwardIntoUnchangeable,
   handleDeleteInsideUnchangeable,
@@ -38,6 +42,10 @@ const withOverrides = (editor: RichTextEditor) => {
       return;
     }
 
+    if (handleDeleteBackwardInPlaceholder(editor)) {
+      return;
+    }
+
     if (handleDeleteBackwardInUndeletable(editor)) {
       return;
     }
@@ -51,6 +59,10 @@ const withOverrides = (editor: RichTextEditor) => {
 
   editor.deleteForward = (unit: TextUnit) => {
     if (handleDeleteInsideUnchangeable(editor, 'forward', unit)) {
+      return;
+    }
+
+    if (handleDeleteForwardInPlaceholder(editor)) {
       return;
     }
